refactor: app keepers
